crossref: disallow redirects when downloading files
A problematic file is [1]. What happens in `fetch` is that this
document URL is forbidden without a proxy (?) and gets redirected
to the paper webpage. The `doi` downloader doesn't have a preferred
file type, so it just downloads the webpage and presents it to
the user.

With this change, we consider any redirection as an error and
don't download the file.

[1] https://onlinelibrary.wiley.com/doi/full/10.1002/nme.1529